src/requirements.txt
torch
yt-dlp
static-ffmpeg
openai-whisper